Ajuste módulo ADC
Se modifica el modo tomar las muestras del conversor A/D:
-La estructura SENSORES almacenará:
*Temperatura en décimas de grado centígrado.
*Pluviometría en litros por metro cuadrado. Max. 200 l/m2.
*Velocidad del aire en kilómetros por hora. Max. 150 km/h.
*Nivel de la batería en porcentaje.
-Se modifica el tipo de estos valores de DWORD a WORD (2 bytes).

+++ General.h
@@ -138,10 +138,10 @@ typedef union _DWORD_VAL
 //Tipo definido para almacenar las señales meteorológicas que se van a muestrear
 typedef struct
 {
-    DWORD    Pluviometria;
-    DWORD    Temperatura;
-    DWORD    Vel_Aire;
-    DWORD    Nivel_Bateria;
+    WORD    Pluviometria;
+    WORD    Temperatura;
+    WORD    Vel_Aire;
+    WORD    Nivel_Bateria;
 }SENSORES;
 
 #endif
